Place Theme wrapper in own folder & add defaults
