+++ frontend/src/components/App/App.js
@@ -2,15 +2,12 @@ import "./App.css";
 import NavBar from "../NavBar/NavBar";
 import Body from "../Body/Body";
 import Footer from "../Footer/Footer";
-import { createMuiTheme, ThemeProvider } from "@material-ui/core";
+
 import React, { useEffect, useState } from "react";
 import { getAll } from '../../services/book';
 import axios from 'axios';
-const theme = createMuiTheme({
-  typography: {
-    fontSize: 14,
-  }
-});
+import { Theme } from '../../theme-style/materialtheme'
+
 
 function App() {
   
@@ -22,11 +19,11 @@ function App() {
 
 
   return (
-    <ThemeProvider theme={theme}>
+    <Theme>
       <NavBar />
       <Body />
       <Footer />
-    </ThemeProvider>
+    </Theme>
     
   );
 }
